+++ src/hcex/hcex_set_pjl_effect.cpp
@@ -8,7 +8,7 @@
 extern "C" int hcex_pjl_effect;
 extern "C" void tag_iterator_new(tag_iterator *iterator, uint32_t key_group_tag);
 extern "C" int tag_iterator_next(tag_iterator *iterator);
-extern "C" char *tag_get_name(int16_t tag_index);
+extern "C" char *tag_get_name(int tag_index);
 extern "C" char       *strstr(const char *haystack, const char *needle);
 
 extern "C" void hcex_set_pjl_effect(const char *name)
