+++ src/rasterizer_dynamic_vertices_get_type.c
@@ -1,6 +1,6 @@
 #include <stdint.h>
 
-extern int16_t _rasterizer_dynamic_vertices_get_type(int dynamic_vertex_buffer_index);
+extern int _rasterizer_dynamic_vertices_get_type(int dynamic_vertex_buffer_index);
 
 int16_t rasterizer_dynamic_vertices_get_type(int dynamic_vertex_buffer_index)
 {
